+++ spec/rubocop/cop/lint/literal_as_condition_spec.rb
@@ -166,7 +166,7 @@ describe RuboCop::Cop::Lint::LiteralAsCondition do
   it 'registers an offense for case with a primitive array condition' do
     expect_offense(<<-RUBY.strip_indent)
       case [1, 2, [3, 4]]
-           ^^^^^^^^^^^^^^ Literal `[1, 2, [3, 4]]` appeared in a condition.
+           ^^^^^^^^^^^^^^ Literal `[1, 2, [3, 4]]` appeared as a condition.
       when [1, 2, 5] then top
       end
     RUBY